(1) Pythia8 seems to work with MPI on now (at least on my laptop), however, I  don't know what I did to get it working
(2) Pythia6 doesn't work WITHOUT MPI - we get a load
of DIVIDE_BY_ZERO and UNDERFLOW flags popping up.
Maybe it will work on the desktop? Although I doubt it

(3) Fixed the selections for str, mttbar etc - I had
just made a simple mistake in the str loop, and
forgotten to enclose the last analyses in the do loops.

(4) Added the jet-pull observable, or at least
something like it.